backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.3
python-decouple>=3.8
django-extensions>=3.2.0
drf-spectacular>=0.27.0
requests>=2.31.0
Pillow>=12.0.0
reportlab>=4.0.0
qrcode>=7.4.2

